refactor(cypress): sort e2e specs by concern, dedup smoke/
Split the overloaded smoke/ dir (DB-task checks + auth journeys +
real smoke) into the dirs that own each concern, and remove duplicate
coverage. smoke/ now holds only genuine smoke (home nav + a11y, env).

- DB-task checks  -> db/   (cleanup, create/delete user, user-exists,
  setup; db:reset task merged into db/reset.cy.ts)
- auth journeys   -> auth/ (new logout.cy.ts; demo-login specs folded
  into access-control as role-based cases; task-login into login.cy.ts)
- stray invoices-list test -> invoices/list.cy.ts
- collapse auth/signup 4 describes -> 1; trim db/setup 4 cases -> 1
- delete exact dup smoke/db-seed (kept db/seed)

Coverage preserved; only duplicate it() blocks removed.
Verified live: pnpm cy:e2e green (30 passing, 0 failing, 2 pre-existing
skips); pnpm typecheck:cypress + pnpm biome:lint clean.
