+++ cypress/e2e/db/delete-user.cy.ts
