--- cypress/e2e/smoke/login-as-demo-admin.cy.ts
@@ -1,10 +1,0 @@
-import { DASHBOARD_USERS_PATH } from "@cypress/e2e/shared/paths";
-
-describe("Login as Demo Admin", () => {
-	it("visits a path restricted to Admins", () => {
-		cy.loginAsDemoAdmin();
-		// visit the user page
-		cy.visit(DASHBOARD_USERS_PATH);
-		cy.location("pathname").should("eq", DASHBOARD_USERS_PATH);
-	});
-});
