--- cypress/e2e/smoke/login-as-demo-user.cy.ts
@@ -1,13 +1,0 @@
-import {
-	DASHBOARD_PATH,
-	DASHBOARD_USERS_PATH,
-} from "@cypress/e2e/shared/paths";
-
-describe("Login as Demo User", () => {
-	it("visits a path restricted to Admins", () => {
-		cy.loginAsDemoUser();
-		// visit the user page
-		cy.visit(DASHBOARD_USERS_PATH);
-		cy.location("pathname").should("eq", DASHBOARD_PATH);
-	});
-});
