--- cypress/e2e/smoke/logout-via-form.cy.ts
@@ -1,17 +1,0 @@
-import { TEN_SECONDS } from "@cypress/e2e/shared/times";
-import { DEMO_USER } from "@cypress/e2e/shared/users";
-
-describe("Logout via Form", () => {
-	const { username, email, password } = DEMO_USER;
-	const signupCreds = { email, password, username };
-
-	beforeEach(() => {
-		cy.logEnv();
-		cy.task("db:deleteUser", signupCreds.email, { timeout: TEN_SECONDS });
-	});
-
-	it("signs up, then logs out via the sign-out button", () => {
-		cy.signup(signupCreds);
-		cy.logoutViaForm();
-	});
-});
